Move files around, speed up objectives a little by using arrays
