+++ src/quester/net/citizensnpcs/questers/QuesterType.java
@@ -5,6 +5,7 @@ import net.citizensnpcs.npctypes.CitizensNPC;
 import net.citizensnpcs.npctypes.CitizensNPCType;
 import net.citizensnpcs.npctypes.NPCTypeManager;
 import net.citizensnpcs.properties.Properties;
+import net.citizensnpcs.questers.data.QuesterProperties;
 import net.citizensnpcs.questers.listeners.QuesterBlockListen;
 import net.citizensnpcs.questers.listeners.QuesterCitizensListen;
 import net.citizensnpcs.questers.listeners.QuesterEntityListen;
