+++ src/quester/net/citizensnpcs/questers/data/QuestProperties.java
@@ -1,6 +1,7 @@
-package net.citizensnpcs.questers;
+package net.citizensnpcs.questers.data;
 
 import net.citizensnpcs.properties.ConfigurationHandler;
+import net.citizensnpcs.questers.QuestManager;
 import net.citizensnpcs.questers.quests.Quest;
 import net.citizensnpcs.questers.quests.QuestFactory;
