+++ src/quester/net/citizensnpcs/questers/quests/QuestFactory.java
@@ -3,7 +3,7 @@ package net.citizensnpcs.questers.quests;
 import net.citizensnpcs.properties.ConfigurationHandler;
 import net.citizensnpcs.properties.Storage;
 import net.citizensnpcs.questers.QuestManager;
-import net.citizensnpcs.questers.Reward;
+import net.citizensnpcs.questers.rewards.Reward;
 import net.citizensnpcs.utils.LocationUtils;
 import net.citizensnpcs.utils.Messaging;
