+++ src/quester/net/citizensnpcs/questers/quests/types/BuildQuest.java
@@ -1,8 +1,8 @@
 package net.citizensnpcs.questers.quests.types;
 
 import net.citizensnpcs.questers.QuestUtils;
-import net.citizensnpcs.questers.quests.ObjectiveProgress;
-import net.citizensnpcs.questers.quests.QuestUpdater;
+import net.citizensnpcs.questers.quests.progress.ObjectiveProgress;
+import net.citizensnpcs.questers.quests.progress.QuestUpdater;
 import net.citizensnpcs.utils.StringUtils;
 
 import org.bukkit.event.Event;
