+++ src/quester/net/citizensnpcs/questers/quests/types/CombatQuest.java
@@ -2,8 +2,8 @@ package net.citizensnpcs.questers.quests.types;
 
 import net.citizensnpcs.PermissionManager;
 import net.citizensnpcs.questers.QuestUtils;
-import net.citizensnpcs.questers.quests.ObjectiveProgress;
-import net.citizensnpcs.questers.quests.QuestUpdater;
+import net.citizensnpcs.questers.quests.progress.ObjectiveProgress;
+import net.citizensnpcs.questers.quests.progress.QuestUpdater;
 
 import org.bukkit.entity.Player;
 import org.bukkit.event.Event;
