+++ src/quester/net/citizensnpcs/questers/quests/types/LocationQuest.java
@@ -1,7 +1,7 @@
 package net.citizensnpcs.questers.quests.types;
 
-import net.citizensnpcs.questers.quests.ObjectiveProgress;
-import net.citizensnpcs.questers.quests.QuestUpdater;
+import net.citizensnpcs.questers.quests.progress.ObjectiveProgress;
+import net.citizensnpcs.questers.quests.progress.QuestUpdater;
 import net.citizensnpcs.utils.LocationUtils;
 import net.citizensnpcs.utils.StringUtils;
