+++ src/quester/net/citizensnpcs/questers/rewards/EconpluginReward.java
@@ -2,8 +2,6 @@ package net.citizensnpcs.questers.rewards;
 
 import net.citizensnpcs.economy.EconomyManager;
 import net.citizensnpcs.properties.Storage;
-import net.citizensnpcs.questers.Reward;
-import net.citizensnpcs.questers.RewardBuilder;
 import net.citizensnpcs.resources.npclib.HumanNPC;
 import net.citizensnpcs.utils.StringUtils;
