+++ src/quester/net/citizensnpcs/questers/rewards/PermissionReward.java
@@ -2,8 +2,6 @@ package net.citizensnpcs.questers.rewards;
 
 import net.citizensnpcs.PermissionManager;
 import net.citizensnpcs.properties.Storage;
-import net.citizensnpcs.questers.Reward;
-import net.citizensnpcs.questers.RewardBuilder;
 import net.citizensnpcs.resources.npclib.HumanNPC;
 
 import org.bukkit.ChatColor;
