+++ src/quester/net/citizensnpcs/questers/rewards/QuestReward.java
@@ -1,10 +1,8 @@
 package net.citizensnpcs.questers.rewards;
 
 import net.citizensnpcs.properties.Storage;
-import net.citizensnpcs.questers.PlayerProfile;
 import net.citizensnpcs.questers.QuestManager;
-import net.citizensnpcs.questers.Reward;
-import net.citizensnpcs.questers.RewardBuilder;
+import net.citizensnpcs.questers.data.PlayerProfile;
 import net.citizensnpcs.resources.npclib.HumanNPC;
 import net.citizensnpcs.utils.StringUtils;
