+++ src/quester/net/citizensnpcs/questers/rewards/Reward.java
@@ -1,4 +1,4 @@
-package net.citizensnpcs.questers;
+package net.citizensnpcs.questers.rewards;
 
 import net.citizensnpcs.properties.Storage;
 import net.citizensnpcs.resources.npclib.HumanNPC;
